Hyrax migration: More fixes
