+++ Gemfile
@@ -94,4 +94,5 @@ group :development, :test do
   gem 'rubocop', '~> 0.42.0'
   # version has to be exactly 1.7, no variation.
   gem 'rubocop-rspec', '1.7'
+  gem 'rails-controller-testing'
 end
