+++ config.ru
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 # This file is used by Rack-based servers to start the application.
 
-require ::File.expand_path('../config/environment', __FILE__)
+require_relative 'config/environment'
 run Rails.application
