+++ config/application.rb
@@ -28,7 +28,7 @@ module ScholarUc
     config.autoload_paths += %W(#{config.root}/app/presenters/concerns)
 
     # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    #config.active_record.raise_in_transactional_callbacks = true
 
     config.exceptions_app = self.routes
