+++ spec/rails_helper.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV['RAILS_ENV'] ||= 'test'
-require File.expand_path('../../config/environment', __FILE__)
+require_relative '../config/environment'
 
 require 'simplecov'
 require 'coveralls'
@@ -132,7 +132,7 @@ RSpec.configure do |config|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
-  config.include Devise::TestHelpers, type: :controller
+  config.include Devise::Test::ControllerHelpers
 
   # Resize the browser window large enough that Capybara can see all elements
   config.before(:each, js: true) do
