+++ test/test_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 ENV['RAILS_ENV'] ||= 'test'
-require File.expand_path('../../config/environment', __FILE__)
+require_relative '../config/environment'
 require 'rails/test_help'
 
 class ActiveSupport::TestCase
